+++ src/ngrams/__init__.py
@@ -1,0 +1,1 @@
+from . ngrams import run
