+++ src/tdm/__init__.py
@@ -1,0 +1,3 @@
+from . entropy  import TermDocumentMatrix
+from . __main__ import get_results
+from . entropy  import n_grams
