--- src/termdocumentmatrix/__init__.py
@@ -1,1 +1,0 @@
-from . termdocumentmatrix import run
