--- src/termdocumentmatrix/termdocumentmatrix.py
@@ -1,308 +1,0 @@
-# encoding: utf-8
-
-import collections
-import unicodedata
-import warnings
-import logging
-import itertools
-import functools
-import string
-import json
-import re
-
-warnings.simplefilter('ignore')
-
-import pandas as pd
-import numpy  as np
-
-from scipy           import sparse, io
-from elasticsearch   import Elasticsearch
-
-from .. environment import cfg
-from .. utils       import search
-
-logging.basicConfig()
-log = logging.getLogger('linkalytics.termdocumentmatrix')
-
-url = cfg["cdr_elastic_search"]["hosts"] + cfg["cdr_elastic_search"]["index"]
-es  = Elasticsearch(url, port=443, verify_certs=False, use_ssl=False, request_timeout=160)
-
-@search(es)
-def get_results(search_term, size, phrase=True):
-    match_type = 'match_phrase' if phrase else 'match'
-    payload = {
-        "size": size,
-        "query" : {
-            match_type : {
-                "_all" : search_term
-            }
-        }
-    }
-    return payload
-
-
-def n_grams(document, n, normalize=True, numbers=True):
-    numbers = 'zero', 'one', 'two', 'three', 'four', 'five', 'six', 'seven', 'eight', 'nine'
-    table   = dict((ord(char), None) for char in string.punctuation)
-    raw     = re.sub('<[^<]+?>', '', document).lower().translate(table)
-
-    # Replace spelled out numbers with actual numbers
-    if numbers:
-        for word, num in list(zip(numbers, range(10))):
-            raw = raw.replace(word, str(num))
-    if normalize:
-        raw     = ''.join(c for c in raw if not unicodedata.combining(c))
-
-    grams = [
-        itertools.islice(raw.split(), i, None) for i in range(n)
-    ]
-    return [' '.join(i) for i in zip(*grams)]
-
-def high_entropy_featurizing(document):
-    """
-    Function to identify attributes of a string to determine whether it may
-    be a unique identifier... WORK IN PROGRESS
-    """
-
-    words = document.split()
-    unique = set()
-    for word in words:
-        n_character = len(word)
-        n_numeric = sum(1 for c in word if c.isnumeric())
-        # n_alpha = sum(1 for c in word if c.isalpha())
-        # n_uppercase = sum(1 for c in word if c.isupper())
-        # n_lowercase = sum(1 for c in word if c.isupper())
-        if (n_character > 8) & (n_character == n_numeric):
-            unique.add(word)
-        # if (...)
-    print(unique)
-
-
-class TermDocumentMatrix:
-    """
-    Efficiently create a term-document matrix.
-    """
-
-    def __init__(self, cutoff=2, tokenizer=n_grams):
-        """
-        :param cutoff: int
-            Specifies only words which appear in minimum documents to be
-            written out as columns in the matrix.
-
-        :param tokenizer: function
-            Function that takes a single string representing a document
-            and return a list of strings representing the n-grams in the document.
-        """
-        self.cutoff    = cutoff
-        self.tokenizer = tokenizer
-        self.sparse    = {}
-
-    def __repr__(self):
-        return '{classname}(cutoff={cutoff}, tokenizer={tokenizer})'.format(
-            classname=self.__class__.__name__,
-            tokenizer=self.tokenizer.__name__,
-            cutoff=self.cutoff,
-        )
-
-    def __len__(self):
-        return len(self.sparse)
-
-    def __iter__(self):
-        """
-        Iterating over this object will output a tuple of key, value pairs
-        """
-        for k, v in self.sparse.items():
-            yield k, v
-
-    def add_doc(self, key, document, n=2, remove_duplicates=True):
-        """
-        Add document to the term-document matrix
-
-        :param document: str
-            String to be tokenized
-
-        :param n: int
-            N-Grams to split using the tokenizer
-
-        :param remove_duplicates: bool
-            Flag that sets whether to add duplicate entries
-        """
-        counter = collections.Counter(self.tokenizer(document, n))
-        cutoff  = {
-            k: 1 for k, v in counter.items()
-                if v >= self.cutoff
-        }
-        if remove_duplicates and cutoff in self.sparse.values():
-            return
-
-        if cutoff:
-            self.sparse[key] = cutoff
-
-    def load_json(self, filepath, *args, **kwargs):
-        """
-        Batch load documents from a fully qualified JSON file.
-
-        :param filepath: str
-            File directory path
-
-        Also takes any arguments allowed by
-
-            TermDocumentMatrix.add_doc(*args, **kwargs)
-
-        """
-        loaded = json.load(open(filepath))
-
-        if isinstance(loaded, list):
-            self.load_list(loaded, *args, **kwargs)
-        else:
-            self.load_dict(loaded, *args, **kwargs)
-
-    def load_dict(self, loaded, *args, **kwargs):
-        """
-        :param loaded: dict
-            Dictionary with the following schema
-
-        :param n: int
-            N-Grams to split using the tokenizer
-
-        Schema
-        ------
-        {
-            “id1”: ”text1”,
-            "id2": "text2"
-        }
-        """
-        for key, document in loaded.items():
-            if isinstance(document, dict) and 'text' in document:
-                self.add_doc(key, document.get('text', None), *args, **kwargs)
-            elif isinstance(document, str):
-                self.add_doc(key, document, *args, **kwargs)
-
-    def load_list(self, loaded, *args, **kwargs):
-        """
-        :param loaded: list
-            Dictionary with the following schema
-
-        :param n: int
-            N-Grams to split using the tokenizer
-
-        Schema
-        ------
-        [
-            {
-                "id": 2314134,
-                "text": "Some Text"
-            },
-            {
-                "id":  4324353,
-                "text" "Some other text"
-            }
-        ]
-        """
-        self.load_dict(
-            {
-                str(item['id']): item['text']
-                    for item in loaded if item.get('text', None)
-            }, *args, **kwargs
-        )
-
-    def to_df(self):
-        """
-        Convert internal TDM representation into a tabular
-        Pandas DataFrame object.
-
-        :return: df
-        :rtype: pd.DataFrame
-
-        """
-        return pd.DataFrame.from_dict(self.sparse, orient='index')\
-                           .fillna(value=0)\
-                           .astype(np.uint16)\
-                           .sort(axis=0, inplace=False)
-
-    def to_sparse(self):
-        """
-        Get the SparseDataFrame representation
-        """
-        df = self.to_df()
-        return sparse.csr_matrix(pd.DataFrame.as_matrix(df), dtype=int), df.index, df.columns
-
-    def save_sparse(self, filename):
-        matrix, index, columns = self.to_sparse()
-        name, _ = filename.rsplit('.')
-
-        io.mmwrite(filename, matrix)
-
-        index.to_series().to_csv('.'.join([name, 'index']))
-        columns.to_series().to_csv('.'.join([name, 'columns']))
-
-    @classmethod
-    def load_sparse(cls, filename):
-        matrix  = io.mmread(filename)
-        name, _ = filename.rsplit('.')
-
-        index   = pd.Series.from_csv('.'.join([name, 'index']))
-        columns = pd.Series.from_csv('.'.join([name, 'columns']))
-
-        return pd.DataFrame(matrix.todense(), index=index, columns=columns).to_sparse(fill_value=0)
-
-    def term2doc(self):
-        """
-        For every term get the documents associated with the term.
-
-        :return: docs
-        :rtype: dict
-        """
-        grams = set()
-
-        for i in self.sparse.values():
-            grams.update(i.keys())
-
-        docs = dict.fromkeys(grams)
-
-        for k, v in self.sparse.items():
-            for word in v.keys():
-                if not docs[word]:
-                    docs[word] = [k]
-                else:
-                    docs[word].append(k)
-
-        return docs
-
-    def sum_columns(self):
-        """
-        Sum up all the table columns
-
-        :return: columns
-        :rtype: pd.Series
-        """
-        c = collections.Counter()
-
-        for key in self.sparse:
-            c.update(self.sparse[key])
-
-        sums = pd.Series(c).sort(inplace=False, ascending=False)
-
-        return sums[sums > self.cutoff]
-
-
-    def write_csv(self, filename):
-        """
-        Write term-document matrix to a CSV file.
-
-        :param filename: Name of the output file (e.g. `mymatrix.csv`).
-        :type  filename: str
-
-        """
-        self.to_df().to_csv(filename, chunksize=128)
-
-
-def run(node):
-    tokenizer = functools.partial(n_grams, numbers=True, normalize=True)
-    tdm       = TermDocumentMatrix(cutoff=1, tokenizer=tokenizer)
-    results   = get_results(node.get('text', None), node.get('size', 1000), True)
-
-    tdm.load_dict(results, n=node.get('ngrams', 2), remove_duplicates=True)
-
-    return {"results": tdm.term2doc()}
-
